Amazon: Support for grocery orders (weighed, tip)
Support for multiple pricing formats involving weighed items:
    * 1 of: 365 Everyday Value, Potato Yellow Bag Organic, 48 Ounce
    * 2 (1.04 lb) of: Broccoli Crowns Conventional, 1 Each
    * 2.07 lb of: Pork Sausage Link Italian Mild Step 1

And ensure the final price for the item is the amount Amazon charged
and not the amount charged multiplied by the "quantity" weighed. This
labels each weighed item entry as a single quantity.

Also add support for parsing out tip left for delivery driver.